payment/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
gevent>=1.4
grpcio==1.64.0
grpcio-tools==1.64.0
msgspec==0.18.6
redis==5.0.3
